+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This is an alpha version! The changes listed here are not final.
 - Social | Fix publicize error in the editor due to malformed connections data
 
 ### Other changes <!-- Non-user-facing changes go here. This section will not be copied to readme.txt. -->
+- Deprecate Jetpack geo location module
 - Fix the rendering of jetpack google fonts font faces for classic themes.
 
 ## 14.3-a.3 - 2025-01-20